% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdset1.R
\name{bset1}
\alias{bset1}
\title{Bounding set for treatment effect with user supplied parameters}
\usage{
bset1(
  beta0,
  R0,
  betatilde,
  Rtilde,
  sigmay,
  sigmax,
  taux,
  N = 100,
  Rlow = 1.3,
  Rhigh = 2.47,
  deltalow,
  deltahigh
)
}
\arguments{
\item{beta0}{treatment effect in short regression}

\item{R0}{R-squared in short regression}

\item{betatilde}{treatment effect in intermediate regression}

\item{Rtilde}{R-squared in intermediate regression}

\item{sigmay}{standard deviation of outcome variable}

\item{sigmax}{standard deviation of treatment variable}

\item{taux}{variance of residual in auxiliary regression (treatment regressed on all included controls)}

\item{N}{used to contruct the grid on which the cubic equation is solved; the grid is N*N; default is N=100}

\item{Rlow}{Rmax in low regime is Rlow*Rtilde; default is Rlow=1.30}

\item{Rhigh}{Rmax in low regime is Rlow*Rtilde; default is Rlow=2.47}

\item{deltalow}{lower bound of delta; this real number must be strictly less than 1}

\item{deltahigh}{lower bound of delta; this real number must be strictly greater than 1}
}
\value{
bset1 returns a list with two elements; the first element is a dataframe containing the parameters supplied by the user about the three regressions; the second element is a matrix containing the quantiles of the empirical distribution of the bias-adjusted treatment effect
}
\description{
This funcition computes quantiles of the empirical distribution of the treatment effect using parameters about the three regressions (short, intermediate and auxiliary) supplied by user.
}
\examples{
# Generate simulated data
# Parameters
a0 <- 0.5;a1 <- 0.25;a2 <- 0.75;a3 <- 1.0
a4 <- -0.75; a5 <- -0.25
# Regressors
set.seed(999)
x1 <- rnorm(100, mean=0, sd=1);x2 <- rnorm(100, mean=1, sd=2)
x3 <- rnorm(100, mean=2, sd=3);x4 <- rnorm(100, mean=3, sd=4)
x5 <- rep(c(0,1),50)
# Dependent variable
y <- a0 + a1*x1 + a2*x2 + a3*x3 + a4*x4 + a5*x5
# Create data frame
d1 <- data.frame(y=y,x1=x1,x2=x2,x3=x3,x4=x4,x5=x5)
# Run short regression and collect paramters
sreg <- stats::lm(y~x1,data=d1)
beta0 <- sreg$coefficients["x1"]
R0 <- summary(sreg)$r.squared
# Run intermediate regression and collect parameters
ireg <- stats::lm(y~x1+x2+x3,data=d1)
betatilde <- ireg$coefficients["x1"]
Rtilde <- summary(ireg)$r.squared
# Run auxiliary regression and collect variance of residuals
auxreg <- stats::lm(x1~x2+x3,data=d1)
taux <- var(auxreg$residuals, na.rm = TRUE)
# Standard deviation of outcome variable
sigmay <- sd(d1$y, na.rm = TRUE)
# Stadard deviation of treatment variable
sigmax <- sd(d1$x1, na.rm=TRUE)
# Generate bounding set for treatment effect with x4 and x5 omitted
bset1(
beta0 = beta0,
R0 = R0,
betatilde = betatilde,
Rtilde = Rtilde,
sigmay = sigmay,
sigmax = sigmax,
taux = taux,
deltalow = 0.01,
deltahigh=5.00
)
}
\references{
Basu, D. (2021). "Bounding Sets for Treatment Effects with Proportional Selection". Economics Department Working Paper Series. 307. University of Massachusetts Amhers. URL: https://scholarworks.umass.edu/econ_workingpaper/307
}
