% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdset.R
\name{bset2}
\alias{bset2}
\title{Bounding set for treatment effect}
\usage{
bset2(
  data,
  outcome,
  treatment,
  shortreg,
  intreg,
  auxreg,
  N = 100,
  Rlow = 1.3,
  Rhigh = 2.47,
  deltalow,
  deltahigh
)
}
\arguments{
\item{data}{dataframe}

\item{outcome}{name of outcome variable as character, e.g. "y"}

\item{treatment}{name of treatment variable as character, e.g. "x1"}

\item{shortreg}{short regression as formula; e.g. y ~ x1}

\item{intreg}{intermediate regression as formula; e.g. y ~ x1 + x2}

\item{auxreg}{auxiliary regression as formula; e.g. x1 ~ x2}

\item{N}{number of points on the grid is N*N; default is N=100}

\item{Rlow}{Rmax in low regime is Rlow*Rtilde; default is Rlow=1.30}

\item{Rhigh}{Rmax in low regime is Rlow*Rtilde; default is Rlow=2.47}

\item{deltalow}{lower bound of delta; this real number must be strictly less than 1}

\item{deltahigh}{lower bound of delta; this real number must be strictly greater than 1}
}
\value{

}
\description{
Run short, intermediate and auxiliary regressions; collect parameters; and, compute quantiles of the empirical distribution of the treatment effect.
}
\examples{
# Generate simulated data
# Parameters
a0 <- 0.5;a1 <- 0.25;a2 <- 0.75;a3 <- 1.0
a4 <- -0.75; a5 <- -0.25
# Regressors
x1 <- rnorm(100, mean=0, sd=1);x2 <- rnorm(100, mean=1, sd=2)
x3 <- rnorm(100, mean=2, sd=3);x4 <- rnorm(100, mean=3, sd=4)
x5 <- rep(c(0,1),50)
# Dependent variable
y <- a0 + a1*x1 + a2*x2 + a3*x3 + a4*x4 + a5*x5
# Create data frame
d1 <- data.frame(y=y,x1=x1,x2=x2,x3=x3,x4=x4,x5=x5)
# Generate bounding set for treatment effect with x4 and x5 omitted
bset2(data = d1,outcome = "y",treatment = "x1",shortreg = y ~ x1,intreg = y ~ x1 + x2 + x3,auxreg = x1 ~ x2 + x3, deltalow = 0.01, deltahigh=5.00)
}
